LLM_RAG_Live/requirements.txt
chromadb
langchain
openai
python-dotenv
bs4
tiktoken
streamlit
streamlit-chat
langchainhub